backend/django_server/requirements.txt
backend==0.2.4.1
Django==5.0
django-cors-headers==4.3.1
django-crum==0.7.9
djangorestframework==3.14.0
drf_spectacular==0.27.1
environ==1.0
gunicorn==21.2.0
Hypercorn==0.15.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.24.0.post1